replacement rule tile and chunk generation
created a replacement rule tile for the tile pallet. this means that I only need to fill the grid with None and not-None flags, which are used to signify when a tile is created and presented.

created a chunk class to hold a collection of tile types and positions to be applied to a tilemap as and when requested. chunks are used as part of a buffering system. one chunk can be created while showing another chunk.

chunk class also contains generation methods to create chunk patterns, such as tunnels, steps and outlines for example.